% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{inflation_indices}
\alias{inflation_indices}
\title{Brazilian inflation indices}
\usage{
inflation_indices()
}
\value{
Vector character
}
\description{
Vector character of all Brazilian inflation indices supported by the Central Bank of Brazil.
}
\examples{
# inflation_indices()
}
