% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_cdi}
\alias{return_cdi}
\title{CDI (Certificado de depósito interbancário or Interbank Money Market)}
\usage{
return_cdi(start_date, end_date, cdi_pct = 100)
}
\arguments{
\item{start_date}{A Date object.}

\item{end_date}{A Date object.}

\item{cdi_pct}{Rate of CDI index.}
}
\value{
A numeric vector of length 1.
}
\description{
Adjustment return of CDI
}
\examples{
\dontrun{
return_cdi(as.Date("2010-01-01"), as.Date("2017-01-01"))
}
}
