% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_inflation}
\alias{return_inflation}
\title{Inflation return}
\usage{
return_inflation(index, start_date, end_date)
}
\arguments{
\item{index}{A inflation index. Must be one of `names(inflation_indices())`.}

\item{start_date}{A date object.}

\item{end_date}{A date object.}
}
\value{
A numeric vector of length 1.
The adjustament return rate from `start_date` to `end_date`
}
\description{
Returns the adjustment return of a selected Brazilian inflation index.
See data(inflation_indices) for all possible values.
}
\examples{
\dontrun{
return_inflation("IGP-M", as.Date("2010-01-01"), as.Date("2017-01-01"))
}
# To adjust a value:
\dontrun{
100 * return_inflation("IGP-M", as.Date("2010-01-01"), as.Date("2017-01-01"))
}
}
