% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_poupanca}
\alias{return_poupanca}
\title{Poupança (Brazilian Savings Rate)}
\usage{
return_poupanca(start_date, end_date)
}
\arguments{
\item{start_date}{A Date object.}

\item{end_date}{A Date object.}
}
\value{
A numeric vector of length 1.
}
\description{
Adjustment return of Poupança.
}
\examples{
\dontrun{
return_poupanca(as.Date("2010-01-01"), as.Date("2017-01-01"))
}
}
