% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_selic}
\alias{return_selic}
\title{Selic (Brazilian Interest Rate)}
\usage{
return_selic(start_date, end_date)
}
\arguments{
\item{start_date}{A Date object.}

\item{end_date}{A Date object.}
}
\value{
A numeric vector of length 1.
}
\description{
Adjustment return of Selic.
}
\examples{
return_selic(as.Date("2010-01-01"), as.Date("2017-01-01"))
}
