% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_tr}
\alias{return_tr}
\title{Taxa Referencial (TR or Referential Rate) return}
\usage{
return_tr(start_date, end_date)
}
\arguments{
\item{start_date}{A Date object.}

\item{end_date}{A Date object.}
}
\value{
A numeric vector of length 1.
}
\description{
Adjustment return of the Referential Rate.
}
\examples{
\dontrun{
return_tr(as.Date("2010-01-01"), as.Date("2017-01-01"))
}
}
